LPS-182759 Split them all to their own class now
