+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-etc-spring-boot/src/main/java/com/liferay/sample/ObjectAction2RestController.java
@@ -1,0 +1,65 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.sample;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+
+import org.json.JSONObject;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @author Raymond Augé
+ * @author Gregory Amerson
+ * @author Brian Wing Shun Chan
+ */
+@RequestMapping("/object/action/2")
+@RestController
+public class ObjectAction2RestController extends BaseRestController {
+
+	@PostMapping
+	public ResponseEntity<String> postSampleObjectAction2(
+		@AuthenticationPrincipal Jwt jwt, @RequestBody String json) {
+
+		if (_log.isInfoEnabled()) {
+			_log.info("JWT Claims: " + jwt.getClaims());
+			_log.info("JWT ID: " + jwt.getId());
+			_log.info("JWT Subject: " + jwt.getSubject());
+
+			try {
+				JSONObject jsonObject = new JSONObject(json);
+
+				_log.info("\n\n" + jsonObject.toString(4) + "\n");
+			}
+			catch (Exception exception) {
+				_log.error("JSON: " + json, exception);
+			}
+		}
+
+		return new ResponseEntity<>(json, HttpStatus.CREATED);
+	}
+
+	private static final Log _log = LogFactory.getLog(
+		ObjectAction2RestController.class);
+
+}
